Req-024 Fixed all UI Bugs, Implemented Row headers, tooltip texts, fixed problem with impact calculations and modified parameter usage to improve performance

+++ src/main/java/de/fau/osr/core/vcs/base/CommitFile.java
@@ -17,7 +17,7 @@ public class CommitFile {
 	public File newPath;
 	public String commitID;
 	public String changedData;
-	public int impact;
+	public float impact;
 	
 	/**
 	 * @author Gayathery
